+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - Support open from clipboard & open from file [#4](https://github.com/ahmedash95/rapidsnap/pull/4)
+- Add Implicit color feature [#5](https://github.com/ahmedash95/rapidsnap/pull/5)
 
 # 1.3.1
 - Fix window menu actions in statusbar by hiding the edit actions [#3](https://github.com/ahmedash95/rapidsnap/pull/3)